requirements.txt
flask
pandas
numpy
scipy
requests
python-dotenv
schedule
pydantic
